+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     - プレイ画面のスクリーンショット共有機能を追加
 
 - Misc
+  - `./webapp`
+    - ゲーム一覧レイアウトの改善
   - other
     - Docker Compose で起動停止を繰り返すと Valkey Cluster の起動が不安定になる問題を修正